enable freelist randomization by default with an RNG that doesnt require a syscall
