+++ include/iso_alloc_internal.h
@@ -388,7 +388,8 @@ INTERNAL_HIDDEN uint64_t _iso_alloc_big_zone_mem_usage(void);
 INTERNAL_HIDDEN uint64_t __iso_alloc_big_zone_mem_usage(void);
 INTERNAL_HIDDEN uint64_t _iso_alloc_mem_usage(void);
 INTERNAL_HIDDEN uint64_t __iso_alloc_mem_usage(void);
-INTERNAL_HIDDEN uint64_t rand_uint64(void);
+INTERNAL_HIDDEN INLINE uint64_t us_rand_uint64(uint64_t *seed);
+INTERNAL_HIDDEN INLINE uint64_t rand_uint64(void);
 INTERNAL_HIDDEN uint8_t _iso_alloc_get_mem_tag(void *p, iso_alloc_zone_t *zone);
 INTERNAL_HIDDEN size_t _iso_alloc_print_stats(void);
 INTERNAL_HIDDEN size_t _iso_chunk_size(void *p);
